refactor: restructure application to separate infra layer
