+++ src/infra/auth/auth.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common'
 import { ConfigService } from '@nestjs/config'
 import { JwtModule } from '@nestjs/jwt'
 import { PassportModule } from '@nestjs/passport'
-import { Env } from '@/env'
+import { Env } from '@/infra/env'
 import { JwtStrategy } from './jwt.strategy'
 
 @Module({
