+++ src/infra/auth/jwt.strategy.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common'
 import { ConfigService } from '@nestjs/config'
 import { PassportStrategy } from '@nestjs/passport'
 import { ExtractJwt, Strategy } from 'passport-jwt'
-import { Env } from '@/env'
+import { Env } from '@/infra/env'
 import { z } from 'zod'
 
 const userSchema = z.object({
